move SchedulingQueue to pkg/scheduler/internal/queue
