+++ pkg/scheduler/factory/cache_comparer.go
@@ -25,14 +25,14 @@ import (
 	"k8s.io/apimachinery/pkg/labels"
 	corelisters "k8s.io/client-go/listers/core/v1"
 	schedulercache "k8s.io/kubernetes/pkg/scheduler/cache"
-	"k8s.io/kubernetes/pkg/scheduler/core"
+	internalqueue "k8s.io/kubernetes/pkg/scheduler/internal/queue"
 )
 
 type cacheComparer struct {
 	nodeLister corelisters.NodeLister
 	podLister  corelisters.PodLister
 	cache      schedulercache.Cache
-	podQueue   core.SchedulingQueue
+	podQueue   internalqueue.SchedulingQueue
 
 	compareStrategy
 }
