+++ pkg/scheduler/internal/queue/scheduling_queue_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package core
+package queue
 
 import (
 	"fmt"
@@ -27,6 +27,7 @@ import (
 	"k8s.io/kubernetes/pkg/scheduler/util"
 )
 
+var negPriority, lowPriority, midPriority, highPriority, veryHighPriority = int32(-100), int32(0), int32(100), int32(1000), int32(10000)
 var mediumPriority = (lowPriority + highPriority) / 2
 var highPriorityPod, highPriNominatedPod, medPriorityPod, unschedulablePod = v1.Pod{
 	ObjectMeta: metav1.ObjectMeta{
